refactor: rename useInfinityGetList o useInfinityList
