+++ packages/core/src/hooks/data/useInfinityList.ts
@@ -38,7 +38,7 @@ export type UseInfinityListProps<TData, TError> = {
     resource: string;
     /**
      * Configuration for pagination, sorting and filtering
-     * @type [`UseInfinityConfig`](/docs/api-reference/core/hooks/data/useInfinityGetList/#config-parameters)
+     * @type [`UseInfinityConfig`](/docs/api-reference/core/hooks/data/useInfinityList/#config-parameters)
      */
     config?: UseInfinityConfig;
     /**
@@ -60,17 +60,17 @@ export type UseInfinityListProps<TData, TError> = {
     LiveModeProps;
 
 /**
- * `useInfinityGetList` is a modified version of `react-query`'s {@link https://tanstack.com/query/latest/docs/react/guides/infinite-queries `useInfiniteQuery`} used for retrieving items from a `resource` with pagination, sort, and filter configurations.
+ * `useInfinityList` is a modified version of `react-query`'s {@link https://tanstack.com/query/latest/docs/react/guides/infinite-queries `useInfiniteQuery`} used for retrieving items from a `resource` with pagination, sort, and filter configurations.
  *
  * It uses the `getList` method as the query function from the `dataProvider` which is passed to `<Refine>`.
  *
- * @see {@link https://refine.dev/docs/core/hooks/data/useInfinityGetList} for more details.
+ * @see {@link https://refine.dev/docs/core/hooks/data/useInfinityList} for more details.
  *
  * @typeParam TData - Result data of the query extends {@link https://refine.dev/docs/core/interfaceReferences#baserecord `BaseRecord`}
  * @typeParam TError - Custom error object that extends {@link https://refine.dev/docs/core/interfaceReferences#httperror `HttpError`}
  *
  */
-export const useInfinityGetList = <
+export const useInfinityList = <
     TData extends BaseRecord = BaseRecord,
     TError extends HttpError = HttpError,
 >({
@@ -183,7 +183,7 @@ export const useInfinityGetList = <
                         : errorNotification;
 
                 handleNotification(notificationConfig, {
-                    key: `${resource}-useInfinityGetList-notification`,
+                    key: `${resource}-useInfinityList-notification`,
                     message: translate(
                         "common:notifications.error",
                         { statusCode: err.statusCode },
